Desktop/rehapp-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

SQLAlchemy==2.0.34
psycopg2-binary==2.9.9

python-multipart==0.0.9